# perf: Perform string->datetime conversion once during timeseries insights