This is a crucial insights optimization that deserves its own PR (and it's maximally uncontroversial).

In the current implementation, the reducing function that performs insights on a timeseries-like data (data with exactly one datetime column) performs pairwise analysis on a timeseries column and each data column. For each pair, the timeseries column is converted string->datetime every time. That explains why I'm seeing so much date conversion work and addressed some of it in the previous PRs.

The patch makes sure that the conversion happens early and only once.

In my synthetic benchmark (timeseries table with ~20 data columns, 20k rows displayed on a dashboard, this is what I'm getting before the patch:

```
Time per call: 2.49 s   Alloc per call: 5,036,408,480b
Time per call: 2.31 s   Alloc per call: 5,036,259,688b
Time per call: 2.38 s   Alloc per call: 5,036,259,088b
Time per call: 2.47 s   Alloc per call: 5,036,259,432b
```

After the patch:

```
Time per call: 1.07 s   Alloc per call: 1,982,806,768b
Time per call: 1.11 s   Alloc per call: 1,982,807,480b
Time per call: 1.25 s   Alloc per call: 1,982,808,592b
Time per call: 1.24 s   Alloc per call: 1,982,804,896b
```